change type of IDs and triggeredForCatalogItem to Long

+++ src/main/java/one/talon/model/AchievementStatusEntry.java
@@ -35,7 +35,7 @@ import org.threeten.bp.OffsetDateTime;
 public class AchievementStatusEntry {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -235,7 +235,7 @@ public class AchievementStatusEntry {
   private AchievementProgress currentProgress;
 
 
-  public AchievementStatusEntry id(Integer id) {
+  public AchievementStatusEntry id(Long id) {
     
     this.id = id;
     return this;
@@ -247,12 +247,12 @@ public class AchievementStatusEntry {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
