+++ src/main/java/one/talon/model/ApplicationCIF.java
@@ -32,7 +32,7 @@ import org.threeten.bp.OffsetDateTime;
 public class ApplicationCIF {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -67,7 +67,7 @@ public class ApplicationCIF {
   private Integer applicationId;
 
 
-  public ApplicationCIF id(Integer id) {
+  public ApplicationCIF id(Long id) {
     
     this.id = id;
     return this;
@@ -79,12 +79,12 @@ public class ApplicationCIF {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
