+++ src/main/java/one/talon/model/ApplicationCIFExpression.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class ApplicationCIFExpression {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -57,7 +57,7 @@ public class ApplicationCIFExpression {
   private Integer applicationId;
 
 
-  public ApplicationCIFExpression id(Integer id) {
+  public ApplicationCIFExpression id(Long id) {
     
     this.id = id;
     return this;
@@ -69,12 +69,12 @@ public class ApplicationCIFExpression {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
