+++ src/main/java/one/talon/model/ApplicationEvent.java
@@ -36,7 +36,7 @@ import org.threeten.bp.OffsetDateTime;
 public class ApplicationEvent {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -79,7 +79,7 @@ public class ApplicationEvent {
   private List<RuleFailureReason> ruleFailureReasons = null;
 
 
-  public ApplicationEvent id(Integer id) {
+  public ApplicationEvent id(Long id) {
     
     this.id = id;
     return this;
@@ -91,12 +91,12 @@ public class ApplicationEvent {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
