+++ src/main/java/one/talon/model/AudienceCustomer.java
@@ -37,7 +37,7 @@ import org.threeten.bp.OffsetDateTime;
 public class AudienceCustomer {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -88,7 +88,7 @@ public class AudienceCustomer {
   private List<Integer> connectedAudiences = null;
 
 
-  public AudienceCustomer id(Integer id) {
+  public AudienceCustomer id(Long id) {
     
     this.id = id;
     return this;
@@ -100,12 +100,12 @@ public class AudienceCustomer {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
