+++ src/main/java/one/talon/model/CampaignStoreBudget.java
@@ -35,7 +35,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CampaignStoreBudget {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -54,7 +54,7 @@ public class CampaignStoreBudget {
   private List<CampaignStoreBudgetLimitConfig> limits = new ArrayList<CampaignStoreBudgetLimitConfig>();
 
 
-  public CampaignStoreBudget id(Integer id) {
+  public CampaignStoreBudget id(Long id) {
     
     this.id = id;
     return this;
@@ -66,12 +66,12 @@ public class CampaignStoreBudget {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
