+++ src/main/java/one/talon/model/CustomerSessionV2.java
@@ -40,7 +40,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CustomerSessionV2 {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -170,7 +170,7 @@ public class CustomerSessionV2 {
   private OffsetDateTime updated;
 
 
-  public CustomerSessionV2 id(Integer id) {
+  public CustomerSessionV2 id(Long id) {
     
     this.id = id;
     return this;
@@ -182,12 +182,12 @@ public class CustomerSessionV2 {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
