+++ src/main/java/one/talon/model/IntegrationCoupon.java
@@ -36,7 +36,7 @@ import org.threeten.bp.OffsetDateTime;
 public class IntegrationCoupon {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -127,7 +127,7 @@ public class IntegrationCoupon {
   private Integer profileRedemptionCount;
 
 
-  public IntegrationCoupon id(Integer id) {
+  public IntegrationCoupon id(Long id) {
     
     this.id = id;
     return this;
@@ -139,12 +139,12 @@ public class IntegrationCoupon {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
