+++ src/main/java/one/talon/model/SamlConnection.java
@@ -69,7 +69,7 @@ public class SamlConnection {
 
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -276,7 +276,7 @@ public class SamlConnection {
   }
 
 
-  public SamlConnection id(Integer id) {
+  public SamlConnection id(Long id) {
     
     this.id = id;
     return this;
@@ -288,12 +288,12 @@ public class SamlConnection {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Internal ID of this entity.")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
